Ref #54517 and friends

```
$ RUBY_YJIT_ENABLE=1 STAGE=after ruby visitor.rb
== index ==
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +YJIT +PRISM [arm64-darwin23]
Warming up --------------------------------------
               after    52.600k i/100ms
Calculating -------------------------------------
               after    531.147k (± 1.0%) i/s    (1.88 μs/i) -      2.683M in   5.051104s

Comparison:
              before:   463679.0 i/s
               after:   531146.6 i/s - 1.15x  faster

== show ==
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +YJIT +PRISM [arm64-darwin23]
Warming up --------------------------------------
               after    41.754k i/100ms
Calculating -------------------------------------
               after    413.785k (± 1.4%) i/s    (2.42 μs/i) -      2.088M in   5.046424s

Comparison:
              before:   368692.0 i/s
               after:   413784.6 i/s - 1.12x  faster

== show_nested ==
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +YJIT +PRISM [arm64-darwin23]
Warming up --------------------------------------
               after    35.727k i/100ms
Calculating -------------------------------------
               after    356.375k (± 1.6%) i/s    (2.81 μs/i) -      1.786M in   5.013885s

Comparison:
              before:   322194.4 i/s
               after:   356375.2 i/s - 1.11x  faster
```

cc @byroot